mitm_forwarder/requirements.txt
mitmproxy==10.1.5
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

